Auth and session

+++ MoviesReviewsPlatform/MoviesReviewsPlatform/Endpoints.cs
@@ -1,9 +1,11 @@
 using System.IdentityModel.Tokens.Jwt;
+using System.Net.Http;
 using System.Security.Claims;
 using System.Text.Json;
 using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authorization.Infrastructure;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using MoviesReviewsPlatform.Auth.Model;
 using MoviesReviewsPlatform.Data;
@@ -88,7 +90,7 @@ namespace MoviesReviewsPlatform
             }).WithName("CreateMovie");
 
             // Update movie
-            moviesGroup.MapPut("movies/{movieId}", [Authorize(Roles = PlatformRoles.PlatformUser)] async (int movieId, [Validate] UpdateMovieDto movieDto, 
+            moviesGroup.MapPut("movies/{movieId}", [Authorize] async (int movieId, [Validate] UpdateMovieDto movieDto, 
                 ForumDbContext dbContext, HttpContext httpContent) =>
             {
                 var movie = await dbContext.Movies.FirstOrDefaultAsync(m => m.Id == movieId);
@@ -116,11 +118,16 @@ namespace MoviesReviewsPlatform
             }).WithName("EditMovie");
 
             // Delete movie
-            moviesGroup.MapDelete("movies/{movieId}", async (int movieId, ForumDbContext dbContext) =>
+            moviesGroup.MapDelete("movies/{movieId}", [Authorize] async (int movieId, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var movie = await dbContext.Movies.FirstOrDefaultAsync(m => m.Id == movieId);
                 if (movie == null)
                     return Results.NotFound();
+
+                if (!httpContext.User.IsInRole(PlatformRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != movie.UserId)
+                {
+                    return Results.Forbid();
+                }
 
                 dbContext.Remove(movie);
                 await dbContext.SaveChangesAsync();
@@ -188,7 +195,7 @@ namespace MoviesReviewsPlatform
             }).WithName("GetReview");
 
             // Create a new review
-            reviewsGroup.MapPost("", async (int movieId, [Validate] CreateReviewDto createReviewDto, 
+            reviewsGroup.MapPost("", [Authorize(Roles = PlatformRoles.PlatformUser)] async (int movieId, [Validate] CreateReviewDto createReviewDto, 
                 ForumDbContext dbContext, LinkGenerator linkGenerator, HttpContext httpContext) =>
             {
                 var movie = await dbContext.Movies.FirstOrDefaultAsync(m => m.Id == movieId);
@@ -213,12 +220,17 @@ namespace MoviesReviewsPlatform
             }).WithName("CreateReview");
 
             // Update a review
-            reviewsGroup.MapPut("{reviewId}", async (int movieId, int reviewId, 
-                [Validate] UpdateReviewDto updateReviewDto, ForumDbContext dbContext) =>
+            reviewsGroup.MapPut("{reviewId}", [Authorize] async (int movieId, int reviewId, 
+                [Validate] UpdateReviewDto updateReviewDto, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var review = await dbContext.Reviews.Include(r => r.Movie)
                     .FirstOrDefaultAsync(r => r.Id == reviewId && r.Movie.Id == movieId);
                 if (review == null) return Results.NotFound();
+
+                if (!httpContext.User.IsInRole(PlatformRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != review.UserId)
+                {
+                    return Results.Forbid();
+                }
 
                 review.Text = updateReviewDto.Text;
                 review.Evaluation = updateReviewDto.Evaluation;
@@ -231,11 +243,16 @@ namespace MoviesReviewsPlatform
             }).WithName("UpdateReview");
 
             // Delete a review
-            reviewsGroup.MapDelete("{reviewId}", async (int movieId, int reviewId, ForumDbContext dbContext) =>
+            reviewsGroup.MapDelete("{reviewId}", [Authorize] async (int movieId, int reviewId, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var review = await dbContext.Reviews.Include(r => r.Movie)
                     .FirstOrDefaultAsync(r => r.Id == reviewId && r.Movie.Id == movieId);
                 if (review == null) return Results.NotFound();
+
+                if (!httpContext.User.IsInRole(PlatformRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != review.UserId)
+                {
+                    return Results.Forbid();
+                }
 
                 dbContext.Reviews.Remove(review);
                 await dbContext.SaveChangesAsync();
@@ -300,7 +317,7 @@ namespace MoviesReviewsPlatform
             }).WithName("GetComment");
 
             // Create a new comment
-            commentsGroup.MapPost("", async (int movieId, int reviewId, 
+            commentsGroup.MapPost("", [Authorize(Roles = PlatformRoles.PlatformUser)] async (int movieId, int reviewId, 
                 [Validate] CreateCommentDto createCommentDto, ForumDbContext dbContext, LinkGenerator linkGenerator, 
                 HttpContext httpContext) =>
             {
@@ -326,12 +343,17 @@ namespace MoviesReviewsPlatform
             }).WithName("CreateComment");
 
             // Update a comment
-            commentsGroup.MapPut("{commentId}", async (int movieId, int reviewId, int commentId, 
-                [Validate] UpdateCommentDto updateCommentDto, ForumDbContext dbContext) =>
+            commentsGroup.MapPut("{commentId}", [Authorize] async (int movieId, int reviewId, int commentId, 
+                [Validate] UpdateCommentDto updateCommentDto, ForumDbContext dbContext, HttpContext httpContext) =>
             {
                 var comment = await dbContext.Comments.Include(c => c.Review).ThenInclude(r => r.Movie)
                     .FirstOrDefaultAsync(c => c.Id == commentId && c.Review.Id == reviewId && c.Review.Movie.Id == movieId);
                 if (comment == null) return Results.NotFound();
+
+                if (!httpContext.User.IsInRole(PlatformRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != comment.UserId)
+                {
+                    return Results.Forbid();
+                }
 
                 comment.Text = updateCommentDto.Text;
                 comment.Date = DateTime.UtcNow; // Update timestamp
@@ -343,11 +365,17 @@ namespace MoviesReviewsPlatform
             }).WithName("UpdateComment");
 
             // Delete a comment
-            commentsGroup.MapDelete("{commentId}", async (int movieId, int reviewId, int commentId, ForumDbContext dbContext) =>
+            commentsGroup.MapDelete("{commentId}", [Authorize] async (int movieId, int reviewId, int commentId, ForumDbContext dbContext,
+                HttpContext httpContext) =>
             {
                 var comment = await dbContext.Comments.Include(c => c.Review).ThenInclude(r => r.Movie)
                     .FirstOrDefaultAsync(c => c.Id == commentId && c.Review.Id == reviewId && c.Review.Movie.Id == movieId);
                 if (comment == null) return Results.NotFound();
+
+                if (!httpContext.User.IsInRole(PlatformRoles.Admin) && httpContext.User.FindFirstValue(JwtRegisteredClaimNames.Sub) != comment.UserId)
+                {
+                    return Results.Forbid();
+                }
 
                 dbContext.Comments.Remove(comment);
                 await dbContext.SaveChangesAsync();
